pyproject.toml
[tool.poetry]
name = "django-ninja-example"
version = "0.1.0"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
django-ninja = "^1.1.0"
pyjwt = "^2.8.0"
pydantic = "^2.6.4"
email-validator = "^2.1.1"
uvicorn = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
